Fix incorrect code (after incorrect fix done in r72683)
The old fix from r72683 does not work anymore in GCC 6.2, because it relied on comparing pointers of different types. This time, a cast to a common base class is used to compare the slots for equality.

Prior to r72683, the comparison was flawed anyway (comparing the result of implicit casts to bool).


git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@75292 613412ce-1361-4b7a-93bf-a3c75af1bae4