define auto_ptr / unique_ptr to the other class depending on C++ version (use unique_ptr with c++11, auto_ptr otherwise)
This avoids hundreds of warnings in C++11 and use unique_ptr which is the right thing. We could remove the definition in C++11 when all uses of auto_ptr are removed from the code.

Most uses of auto_ptr have been replaced with unique_ptr (some remain in weak_ptr conversions)



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/branches/bug_fix@79884 613412ce-1361-4b7a-93bf-a3c75af1bae4