Restore writing of dummy orientation qform to the Minf header
This change largely reverts r83111. Some code (notably DISCO) relies on the
scanner-based qform being always present in the Minf file.

This change also refactors the way that the dummy scanner-based orientation
transform is inserted. In particular, the dummy transform will now also end up
in sform if there are no other transformations. This is more consistent with
the behaviour for other (non-dummy) transformations.



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/branches/bug_fix@83204 613412ce-1361-4b7a-93bf-a3c75af1bae4